Move some developer-specific tools from xls/tools to xls/dev_tools. Also move "dump_llvm_artifacts_main" to dev_tools. Some tools still need to be moved over, but this is good for the first batch.
PiperOrigin-RevId: 667783815
