+++ xls/dev_tools/extract_interface_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "xls/tools/extract_interface.h"
+#include "xls/dev_tools/extract_interface.h"
 
 #include <string_view>
