+++ xls/dev_tools/extract_sample_points_from_ir_main.cc
