+++ xls/dev_tools/extract_segment.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "xls/tools/extract_segment.h"
+#include "xls/dev_tools/extract_segment.h"
 
 #include <memory>
 #include <optional>
