+++ xls/dev_tools/extract_stage_main.cc
