+++ xls/dev_tools/ir_stats_main.cc
