+++ xls/dev_tools/package_bazel_build_testdata/bazel-bin/package_test.runfiles/c/copied_dep.sh
