+++ xls/tools/opt_main.cc
@@ -39,10 +39,10 @@
 #include "xls/common/file/filesystem.h"
 #include "xls/common/init_xls.h"
 #include "xls/common/status/status_macros.h"
+#include "xls/dev_tools/tool_timeout.h"
 #include "xls/passes/optimization_pass.h"
 #include "xls/passes/optimization_pass_pipeline.h"
 #include "xls/tools/opt.h"
-#include "xls/tools/tool_timeout.h"
 
 static constexpr std::string_view kUsage = R"(
 Takes in an IR file and produces an IR file that has been run through the
